WIP: Development of Socket Implementation
